+++ Plutarch/Prelude.hs
@@ -254,12 +254,18 @@ module Plutarch.Prelude (
   ptraceInfoShowId,
   ptraceShowId,
 
-  -- * Semigroup and monoid
+  -- * Semigroup and Monoid
   PSemigroup (..),
   PMonoid (..),
   PAnd (..),
   POr (..),
   PXor (..),
+
+  -- * Unrolling utilities
+  punrollBound,
+  punrollBound',
+  punrollUnbound,
+  punrollUnboundWhole,
 ) where
 
 import Plutarch.Builtin
@@ -297,4 +303,5 @@ import Plutarch.Pair
 import Plutarch.Rational
 import Plutarch.TermCont
 import Plutarch.Trace
+import Plutarch.Unroll
 import Prelude ()